+++ example/lib/screens/cart_tab.dart
@@ -29,8 +29,9 @@ class _CartTabState extends State<CartTab> {
   }
 
   void init() {
-    createCart();
-    // getCartById("gid://shopify/Cart/....");
+    // createCart();
+    getCartById(
+        "gid://shopify/Cart/Z2NwLWFzaWEtc291dGhlYXN0MTowMUo2VkZYNk1GWTc0U0NYMDhNRUNSRk5TQw?key=521a54e45a74a00ebbcc36fa425d0f61");
 
     getNProducts();
   }
@@ -270,6 +271,7 @@ class _CartInfoState extends State<CartInfo> {
 
   void onCheckoutTap() async {
     final checkoutUrl = cart.checkoutUrl;
+    log('checkoutUrl: $checkoutUrl');
     if (checkoutUrl == null) {
       context.showSnackBar('Invalid checkout url');
       return;