+++ lib/graphql_operations/storefront/mutations/cart/cart_create.dart
@@ -110,6 +110,12 @@ mutation cartCreate($country: CountryCode, $input: CartInput!) @inContext(countr
           node {
             id
             quantity
+            discountAllocations {
+              discountedAmount {
+                amount
+                currencyCode
+              }
+            }
             sellingPlanAllocation {
               sellingPlan {
                 id